requirements.txt
jwave
jupyterlab
matplotlib
torchvision
tqdm
pre-commit
pytest
flax
optax
git+https://github.com/astanziola/fourier-neural-operator-flax.git # FNO in flax
neptune-client
